Moved the Vc mathematics "under the new organisation".
While also adding a unit test for this setup.
