+++ frontend/CMakeLists.txt
@@ -16,6 +16,7 @@ if( ALGEBRA_PLUGIN_INCLUDE_SMATRIX )
 endif()
 if( ALGEBRA_PLUGIN_INCLUDE_VC )
    add_subdirectory( vc_cmath )
+   add_subdirectory( vc_vc )
 endif()
 if( ALGEBRA_PLUGIN_INCLUDE_VECMEM )
    add_subdirectory( vecmem_cmath )
