+++ frontend/vc_vc/CMakeLists.txt
@@ -1,0 +1,11 @@
+# Algebra plugins library, part of the ACTS project (R&D line)
+#
+# (c) 2021 CERN for the benefit of the ACTS project
+#
+# Mozilla Public License Version 2.0
+
+# Set up the library.
+algebra_add_library( algebra_vc_vc vc_vc
+   "include/algebra/vc_vc.hpp" )
+target_link_libraries( algebra_vc_vc
+   INTERFACE algebra::common algebra::vc_storage algebra::vc_math )
