+++ math/CMakeLists.txt
@@ -12,3 +12,6 @@ endif()
 if( ALGEBRA_PLUGIN_INCLUDE_SMATRIX )
    add_subdirectory( smatrix )
 endif()
+if( ALGEBRA_PLUGIN_INCLUDE_VC )
+   add_subdirectory( vc )
+endif()
