+++ math/vc/include/algebra/math/impl/vc_polar2.hpp
@@ -1,0 +1,59 @@
+/** Algebra plugins, part of the ACTS project
+ *
+ * (c) 2020 CERN for the benefit of the ACTS project
+ *
+ * Mozilla Public License Version 2.0
+ */
+
+#pragma once
+
+// Project include(s).
+#include "algebra/common/algebra_qualifiers.hpp"
+#include "algebra/math/impl/cmath_getter.hpp"
+#include "algebra/storage/vc.hpp"
+
+namespace algebra::vc::math {
+
+/** Local frame projection into a polar coordinate frame */
+template <typename transform3_t>
+struct polar2 {
+
+  /// @name Type definitions for the struct
+  /// @{
+
+  /// Transformation matching this struct
+  using transform3_type = transform3_t;
+
+  /// Point in 2D space
+  using point2 = typename transform3_type::point2;
+  /// Point in 3D space
+  using point3 = typename transform3_type::point3;
+
+  /// @}
+
+  /** This method transform from a point from the global 3D cartesian
+   *  frame to the local 2D cartesian frame
+   *
+   * @param trf the transform from global to local thredimensional frame
+   * @param p the point in global frame
+   *
+   * @return a local point2
+   **/
+  ALGEBRA_HOST_DEVICE
+  inline point2 operator()(const transform3_type &trf, const point3 &p) const {
+
+    return operator()(trf.point_to_local(p));
+  }
+
+  /** This method transform from a point from 2D or 3D cartesian frame
+      to a 2D polar point */
+  template <typename point_type>
+  ALGEBRA_HOST_DEVICE inline point2 operator()(const point_type &v) const {
+
+    return point2{cmath::perp<vc::storage_type>(v),
+                  cmath::phi<vc::storage_type>(v)};
+  }
+
+};  // struct polar2
+
+}  // namespace algebra::vc::math
