+++ math/vc/include/algebra/math/vc.hpp
@@ -1,0 +1,16 @@
+/** Algebra plugins, part of the ACTS project
+ *
+ * (c) 2020 CERN for the benefit of the ACTS project
+ *
+ * Mozilla Public License Version 2.0
+ */
+
+#pragma once
+
+// Project include(s).
+#include "algebra/math/impl/cmath_getter.hpp"
+#include "algebra/math/impl/vc_cartesian2.hpp"
+#include "algebra/math/impl/vc_cylindrical2.hpp"
+#include "algebra/math/impl/vc_polar2.hpp"
+#include "algebra/math/impl/vc_transform3.hpp"
+#include "algebra/math/impl/vc_vector.hpp"
