+++ tests/vc/vc_vc.cpp
@@ -1,0 +1,13 @@
+/** Algebra plugin library, part of the ACTS project
+ *
+ * (c) 2020 CERN for the benefit of the ACTS project
+ *
+ * Mozilla Public License Version 2.0
+ */
+
+// Project include(s).
+#include "algebra/vc_vc.hpp"
+
+// Test include(s).
+#define __plugin algebra::vc
+#include "test_plugin.inl"
